+++ src/evolduo_app/controllers/user.clj
@@ -13,14 +13,14 @@
   [{:keys [session] :as req}]
   (let [captcha (er/random-num 6)]
     (->
-      (r/render-html user-views/signup-form req {:captcha captcha})
+      (r/render-html user-views/signup-form req {:captcha captcha :signup {:newsletters "on"}})
       (assoc :session (assoc session :captcha captcha)))))
 
 (defn signup [req]
   (let [db (:db req)
         captcha-code (or (-> req :session :captcha) "")
         real-ip (req/get-x-forwarded-for-header req)
-        params (-> req :params (select-keys [:email :password :password_confirmation :captcha]))
+        params (-> req :params (select-keys [:email :password :password_confirmation :captcha :newsletters]))
         sanitized-data (s/decode-and-validate s/Signup params)]
     (cond
       (:error sanitized-data)
@@ -53,7 +53,8 @@
         (if-let [user (user/upsert!
                         db
                         (-> sanitized-data :data :email)
-                        (-> sanitized-data :data :password))]
+                        (-> sanitized-data :data :password)
+                        (-> sanitized-data :data :newsletters))]
           (->
             (r/redirect "/"
               :flash {:type :info :message [:span
@@ -70,21 +71,27 @@
   [req]
   (r/render-html user-views/login-form req))
 
-(defn login [req]
-  (let [db (:db req)
-        email (-> req :params :email)
+(defn login [{:keys [db session] :as req}]
+  (let [email (-> req :params :email)
         password (-> req :params :password)
-        real-ip (req/get-x-forwarded-for-header req)
-        session (:session req)]
-    (if-let [user (user/login-user db email password)]
-      (let [session' (assoc session :user/id (:id user))]
-        (-> (response/redirect "/")
-          (assoc :session session')))
+        sanitized-data (s/decode-and-validate s/Login {:email email :password password})
+        real-ip (req/get-x-forwarded-for-header req)]
+    (cond
+      (:error sanitized-data)
       (do
         (log/warn "Invalid login attempt from" real-ip)
         (r/render-html user-views/login-form req {:login        {:email email}
                                                   :notification {:type    "danger"
-                                                                 :message "Invalid email or password"}})))))
+                                                                 :message "Invalid email or password"}}))
+
+      :else
+      (if-let [user (user/login-user db
+                                     (-> sanitized-data :data :email)
+                                     (-> sanitized-data :data :password))]
+        (let [session' (assoc session :user/id (:id user))]
+          (-> (response/redirect "/")
+              (assoc :session session')))
+        ))))
 
 (defn logout-user [req]
   (r/logout {:message "You have successfully been logged out"}))
@@ -159,3 +166,72 @@
         (r/logout {:message "Your account has been deleted"})
         ))
     ))
+
+(defn set-password-form [{:keys [db] :as req}]
+  (let [real-ip (req/get-x-forwarded-for-header req)]
+    (if (user/find-user-by-password-reset-token db (-> req :params :token))
+      (r/render-html user-views/set-password-form req)
+      (do
+        (log/warn "Invalid reset password token provided from" real-ip)
+        (r/redirect "/"
+          :flash {:type :danger :message [:span "Password reset token is invalid or expired"]})))))
+
+(defn set-password [{:keys [db] :as req}]
+  (let [params         (-> req :params (select-keys [:token :password :password_confirmation]))
+        sanitized-data (s/decode-and-validate s/PasswordSet params)
+        {:keys [token password]} (-> sanitized-data :data)]
+    (cond
+      (:error sanitized-data)
+      (r/render-html user-views/set-password-form req {:errors (:error sanitized-data)})
+
+      :else
+      (if-let [user (user/find-user-by-password-reset-token db token)]
+        (do
+          (user/update-user-password! db (:id user) password)
+          (r/redirect "/"
+                      :flash {:type :info :message [:span "Your password has been updated successfully"]}))
+        (r/redirect "/"
+                    :flash {:type :danger :message [:span "Password reset is token invalid or expired"]})))
+      ))
+
+(defn password-reset-save [{:keys [db] :as req}]
+  (let [captcha-code   (or (-> req :session :captcha) "")
+        real-ip        (req/get-x-forwarded-for-header req)
+        params         (-> req :params (select-keys [:email :captcha]))
+        sanitized-data (s/decode-and-validate s/PasswordReset params)]
+    (cond
+      (:error sanitized-data)
+      (r/render-html user-views/password-reset-form req {:captcha captcha-code
+                                                         :reset params
+                                                         :errors (:error sanitized-data)})
+
+      (not= captcha-code (-> sanitized-data :data :captcha))
+      (do
+        (log/warn "Invalid password request attempt from" real-ip)
+        (r/render-html user-views/password-reset-form req {:captcha captcha-code
+                                                           :reset params
+                                                           :errors {:captcha ["not valid"]}}))
+
+      ;; TODO check too many requests and update fail2ban
+
+      :else
+      (do
+        (let [user (user/find-user-by-email
+                     db
+                     (-> sanitized-data :data :email))]
+          (when user
+            (log/info "Password reset request from user" (:id user))
+            (user/update-password-reset-token db (:id user)))
+          (r/redirect "/"
+                      :flash {:type :info :message [:span
+                                                    (str "A password reset email has been sent to "
+                                                         (-> sanitized-data :data :email) ". "
+                                                         "Don't forget to check your spam folder.")]})))
+      ))
+  )
+
+(defn password-reset-form [{:keys [session] :as req}]
+  (let [captcha (er/random-num 6)]
+    (->
+      (r/render-html user-views/password-reset-form req {:captcha captcha})
+      (assoc :session (assoc session :captcha captcha)))))