+++ src/evolduo_app/model/user.clj
@@ -16,7 +16,7 @@
 ;; Not using pepper per https://stackoverflow.com/questions/16891729/best-practices-salting-peppering-passwords
 (defn create
   "Returns only the user id"
-  [db email pass]
+  [db email pass newsletters]
   (let [encrypted (password/encrypt pass)
         verification_token (rnd/hex 100)
         unsubscribe_token (rnd/hex 100)]
@@ -27,7 +27,7 @@
                      :verified           false
                      :verification_token verification_token
                      :subscription       {:notifications true
-                                          :announcements true}
+                                          :announcements (some? newsletters)}
                      :unsubscribe_token  unsubscribe_token})))
 
 (defn create-stub
@@ -64,10 +64,11 @@
                 :verified
                 :verification_token
                 :subscription
-                :unsubscribe_token])
+                :unsubscribe_token
+                :password_reset_token])
 
 (defn upsert!
-  [db email pass]
+  [db email pass newsletters]
   (jdbc/with-transaction [tx db]
     (let [tx-opts (jdbc/with-options tx {:builder-fn rs/as-unqualified-lower-maps})]
       (let [user (if-let [user (find-user-by-email tx-opts email)]
@@ -82,11 +83,11 @@
                         :verified           false
                         :verification_token verification_token   ;; TODO duplicated
                         :subscription       {:notifications true ;; TODO hm
-                                             :announcements true}
+                                             :announcements (some? newsletters)}
                         :unsubscribe_token unsubscribe_token}
                        {:email email})
                      user)
-                   (create tx-opts email pass))]
+                   (create tx-opts email pass newsletters))]
         (mail-model/insert tx-opts {:recipient_id (:id user)
                                     :type         "signup"
                                     :data         {}})
@@ -96,6 +97,7 @@
   (let [db (:database.sql/connection integrant.repl.state/system)]
     (get-registered-user db "bar@example.com")))
 
+;; TODO verify usages (create verified?(user-id) function), remove unused calls
 (defn find-user-by-id
   [db id]
   (-> (sql/get-by-id db :users id)
@@ -152,3 +154,34 @@
                                          :subscription {}}
                    {:id user-id})]
     res))
+
+(defn update-password-reset-token
+  "Updates the token and sends an email to the user"
+  [db user-id]
+  (when user-id
+    (let [token (rnd/hex 100)]
+      (jdbc/with-transaction [tx db]
+        (let [tx-opts (jdbc/with-options tx {:builder-fn rs/as-unqualified-lower-maps})]
+          (sql/update! tx-opts :users {:password_reset_token   token
+                                       :password_reset_sent_at (Instant/now)}
+                       {:id user-id})
+          (mail-model/insert tx-opts {:recipient_id user-id
+                                      :type         "password-reset"
+                                      :data         {}}))))))
+
+(defn find-user-by-password-reset-token [db token]
+  (first (sql/find-by-keys db :users {:password_reset_token token})))
+
+(defn update-user-password! [db user-id password]
+  (let [encrypted (password/encrypt password)]
+    (sql/update! db :users {:password encrypted
+                            :password_reset_token nil
+                            :password_reset_sent_at nil} {:id user-id})))
+
+(defn find-users-with-enabled-notifications [db]
+  (map #(select-keys % user-keys)
+       (sql/query db ["select * from users where (cast (subscription ->> 'notifications' as boolean) is true)"])))
+
+(comment
+  (let [db (:database.sql/connection integrant.repl.state/system)]
+    (find-users-with-enabled-notifications db)))